Conflitos de mudança na assinatura, resolvidos.

+++ projetao/safebox/views.py
@@ -105,6 +105,10 @@ def cliente_assinatura_view(request, email):
 
     if action_remover == 'Remover Plano':
         # messages.info(request, "Assinatura desfeita com sucesso!")
+        ambientes = Ambiente.objects.all().filter(cliente_id=context['data'].id)
+        for amb in ambientes:
+            amb.delete()
+
         remover_assinatura(request, email)
         return redirect('visualizar', email)
 
@@ -255,13 +259,10 @@ def assinatura_trocar_view(request, email):
     context['data'] = Cliente.objects.get(email=email)
     assinaturas = Assinatura.objects.all()
     assinatura = assinaturas.filter(cliente_id=context['data'].id)
-    ambientes = Ambiente.objects.all().filter(client_id=)
+    ambientes = Ambiente.objects.all()
+    cameras = Camera.objects.all()
+    planos = {1:1, 2:3, 3:999}
 
-    #####################################################################
-    # Verifica o número de câmeras no plano do cliente
-    planos = Plano.objects.all()
-    plano = planos.filter(nome=assinatura.get_plano_id())[0]
-    max_cams = plano.get_num_cam()
 
     #####################################################################
     # Pega o número de câmeras atuais do cliente
@@ -278,14 +279,17 @@ def assinatura_trocar_view(request, email):
     if (assinatura != None) and (len(assinatura) != 0):
         flag_assinatura_existente = True
 
-
-
-
     form = AssinaturaForm(request.POST or None, initial={"cliente_id": context['data'].id})
+
     if form.is_valid() and flag_assinatura_existente:
-        assinatura.delete()
-        form.save()
-        return redirect('visualizar_assinatura', email)
+
+        max_cams = planos[int(form['plano_id'].value())]
+        if numero_de_cameras_atuais <= max_cams:
+            assinatura.delete()
+            form.save()
+            return redirect('visualizar_assinatura', email)
+        else:
+            messages.success(request, 'A troca para este plano requer que você tenha um número de câmeras compativel. Delete as câmeras que você não deseja mais usar.')
 
     planos = Plano.objects.all()
     context['form'] = form
@@ -654,7 +658,10 @@ def camera_create_view(request, email, nome):
     #####################################################################
     # Verifica o número de câmeras no plano do cliente
 
-    Assin = Assinatura.objects.get(cliente_id=context['data'].id)
+    try:
+        Assin = Assinatura.objects.get(cliente_id=context['data'].id)
+    except:
+        return redirect('home', email)
     planos = Plano.objects.all()
     plano = planos.filter(nome=Assin.get_plano_id())[0]
     max_cams = plano.get_num_cam()
